Added new algorithm for docking

+++ Classes/BoatClasses/Boat.cs
@@ -9,7 +9,7 @@ namespace Hamnen.Classes
         public string Id { get; set; }
         public string DockedAt { get; set; }
         public string BoatType { get; set; }
-        public float Size { get; set; }
+        public int Size { get; set; }
         public int Weight { get; set; }
         public double MaxVelocity { get; set; }
         public int TimeBeforeLeaving { get; set; }
